Move non unit test out from the package for unit tests
